scaffold: reorganize components structure
