+++ README.md
@@ -1,8 +1,8 @@
-# @e-toast/react - The React Layer
+# 📘 @e-toast/react - The React Layer
 
 This contains the React implementation of e-toast.
 
-## Instalation
+## 🏗️ Instalation
 
 *IMPORTANT* this project uses `yarn` version `1.22.1` as package manager. Please avoid using any other package manager.
 
@@ -10,7 +10,24 @@ This contains the React implementation of e-toast.
 yarn add @e-toast/react
 ```
 
-## Usage
+## 🧰 Tech Stack
+
+- React v17.0.0
+- Storybook v.6.2.5
+- Typescript v.4.1.2
+- Jest v27.4.3
+
+## 📁 Folder Structure
+Please follow the [guidelines](https://twitter.com/housecor/status/1493947989877997571?s=20&t=47Knp8NaM3KOzADGxkTH0w) suggested by Cory House:
+
+1. Button.tsx - component
+2. Button.types.ts - TS types
+3. Button.test.ts - Jest + testing-library unit tests
+4. Button.stories.tsx - Storybook stories
+5. ~Button.module.scss - CSS module~ (Replaced by e-toast/css)
+6. index.tsx - Re-exports ([barrel](https://trusz.github.io/posts/barrel-pattern-in-typescript/))
+
+## ✍️ Usage
 
 Components are built using typescript. Can be used both in JS and TS.
 
@@ -25,7 +42,7 @@ import Button from '@e-toast/lib/Button';
 This library should be used with an implementation of ```@e-toast/css``` which needs to be imported before any component (normally in the entry point of your react application). Otherwise the components will have no style applied.
 
 
-## Unit Testing 
+## 🧪 Unit Testing
 
 This repo uses different tools for performing unit test in each component. 
 - [Jest](https://jestjs.io/)
@@ -33,7 +50,7 @@ This repo uses different tools for performing unit test in each component.
 - [testing-library/jest-dom](https://testing-library.com/docs/ecosystem-jest-dom)
 - [testing-library/react](https://testing-library.com/docs/react-testing-library/intro)
 - [testing-library/user-event](https://testing-library.com/docs/ecosystem-user-event/)
-### Running all the tests 
+### Running all the tests
 ```bash
 yarn test
 ```
