+++ package.json
@@ -23,35 +23,35 @@
     "build-storybook": "build-storybook"
   },
   "dependencies": {
-    "@e-toast/react": "^0.1.0-alpha",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "getclassname": "^1.0.9",
-    "typescript": "^4.1.2"
+    "@e-toast/react": "~0.1.0-alpha",
+    "@types/node": "~12.0.0",
+    "@types/react": "~17.0.0",
+    "@types/react-dom": "~17.0.0",
+    "getclassname": "~1.0.9"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/preset-env": "^7.16.4",
-    "@babel/preset-typescript": "^7.16.0",
+    "@babel/core": "~7.16.0",
+    "@babel/preset-env": "~7.16.4",
+    "@babel/preset-typescript": "~7.16.0",
     "@e-toast/css": "latest",
-    "@storybook/addon-actions": "^6.2.5",
-    "@storybook/addon-essentials": "^6.2.5",
-    "@storybook/addon-links": "^6.2.5",
-    "@storybook/react": "^6.2.5",
-    "@testing-library/dom": "^8.11.1",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.0.3",
-    "babel-jest": "^27.4.2",
-    "babel-loader": "^8.2.2",
-    "jest": "^27.4.3",
-    "plop": "^2.7.4",
-    "rimraf": "^3.0.2"
+    "@storybook/addon-actions": "~6.2.5",
+    "@storybook/addon-essentials": "~6.2.5",
+    "@storybook/addon-links": "~6.2.5",
+    "@storybook/react": "~6.2.5",
+    "@testing-library/dom": "~8.11.1",
+    "@testing-library/jest-dom": "~5.16.1",
+    "@testing-library/react": "~12.1.2",
+    "@testing-library/user-event": "~13.5.0",
+    "@types/jest": "~27.0.3",
+    "babel-jest": "~27.4.2",
+    "babel-loader": "~8.2.2",
+    "jest": "~27.4.3",
+    "plop": "~2.7.4",
+    "rimraf": "~3.0.2",
+    "typescript": "~4.1.2"
   },
   "peerDependencies": {
-    "react": "^17.0.2",
+    "react": "~17.0.2",
     "react-dom": "17.0.2"
   },
   "jest": {
