+++ src/Button/Button.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from './index'
+import Button from './Button'
 import { render } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
