+++ src/Select/Select.stories.tsx
@@ -1,30 +1,30 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react';
-import Select, { SelectProps } from '.';
+import Select, { SelectProps } from './Select';
 import {
-    objectControl,
-    noControl,
-    booleanControl,
-    textControl
-}from '../../stories/utils'
-  
-  export default {
-    title: '@e-toast/Select',
-    component: Select,
-    argTypes: {
-      options: objectControl,
-      onChange: noControl(),
-      disabled: booleanControl(),
-      selected: textControl(),    
-    }
-  } as Meta;  
+  objectControl,
+  noControl,
+  booleanControl,
+  textControl
+} from '../../stories/utils'
 
-const Template: Story<SelectProps> = (args) => <Select {...args}/>
+export default {
+  title: '@e-toast/Select',
+  component: Select,
+  argTypes: {
+    options: objectControl,
+    onChange: noControl(),
+    disabled: booleanControl(),
+    selected: textControl(),
+  }
+} as Meta;
+
+const Template: Story<SelectProps> = (args) => <Select {...args} />
 
 export const Controlled = Template.bind({})
 Controlled.args = {
   selected: 'Medium',
-  options: ['Small', 'Medium', 'Large'] 
+  options: ['Small', 'Medium', 'Large']
 }
 
 export const Uncontrolled = Template.bind({})
