+++ src/Select/option.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-interface data{
-  value:string
+interface data {
+  value: string
 }
 
 export interface OptionProps {
@@ -10,22 +10,21 @@ export interface OptionProps {
   onSelected: React.ChangeEventHandler<HTMLDivElement>
 }
 
-const Option: React.FC<OptionProps> = ({options, display, onSelected}) => {
+const Option: React.FC<OptionProps> = ({ options, display, onSelected }) => {
 
-  const onClickOption = (event) => {   
-    onSelected(event);        
+  const onClickOption = (event) => {
+    onSelected(event);
   }
-  
-  return(
-    <div className={display ? 'eds-select__list eds-select__list--open' : 'eds-select__list'}>     
-      {   
-        options.map((datos:data, index:number) =>
+
+  return (
+    <div className={display ? 'eds-select__list eds-select__list--open' : 'eds-select__list'}>
+      {
+        options.map((datos: data, index: number) =>
           <div key={index} className='eds-select__option' data-index={index} onClick={onClickOption}>{datos}</div>
         )
-      }      
-    </div>    
-  ) 
+      }
+    </div>
+  )
 }
 
 export default Option;
-
