+++ src/TextArea/TextArea.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, Story } from '@storybook/react'
-import TextArea, { TextAreaProps } from '.'
-import { booleanControl, noControl, textControl } from '../../stories/utils'
+import TextArea, { TextAreaProps } from './TextArea'
+import { booleanControl, textControl } from '../../stories/utils'
 
 export default {
   title: "@e-toast/TextArea",
