+++ src/index.ts
@@ -1,14 +1,13 @@
-export { default as Avatar } from './Avatar';
-export { default as Button } from './Button';
-export { default as Checkbox } from './Checkbox';
-export { default as Datepicker } from './Datepicker';
-export { default as FieldContainer } from './FieldContainer';
-export { default as FlexLayout } from './FlexLayout';
-export { default as Form } from './Form';
-export { default as NativeDatepicker } from './NativeDatepicker';
-export { default as Select } from './Select';
-export { default as Tag } from './Tag';
-export { default as TextArea } from './TextArea';
-export { default as TextField } from './TextField';
-export { default as PasswordField } from './PasswordField';
-export { default as Toggle } from "./Toggle";
+export { default as Avatar } from './Avatar/Avatar';
+export { default as Button } from './Button/Button';
+export { default as Checkbox } from './Checkbox/Checkbox';
+export { default as Datepicker } from './Datepicker/Datepicker';
+export { default as FieldContainer } from './FieldContainer/FieldContainer';
+export { default as FlexLayout } from './FlexLayout/FlexLayout';
+export { default as Form } from './Form/Form';
+export { default as NativeDatepicker } from './NativeDatepicker/NativeDatepicker';
+export { default as Select } from './Select/Select';
+export { default as Tag } from './Tag/Tag';
+export { default as TextArea } from './TextArea/TextArea';
+export { default as TextField } from './TextField/TextField';
+export { default as Toggle } from "./Toggle/Toggle";
