+++ stories/types.ts
@@ -1,60 +1,60 @@
 export type Sizes = "small"
-export type Icons = 
-| "chevron"  
-| "arrow" 
-| "warning-circle"
-| "warning-circle-fill"
-| "information"
-| "clock"
-| "warning"
-| "notification"
-| "upload"
-| "new-folder"
-| "cloud"
-| "copy"
-| "document"
-| "folder"
-| "new-file"
-| "arrow-down"
-| "arrow-left"
-| "arrow-right"
-| "arrow-up"
-| "back"
-| "caret-down"
-| "caret-left"
-| "caret-right"
-| "caret-up"
-| "down"
-| "full-screen"
-| "menu"
-| "more"
-| "next"
-| "search"
-| "settings"
-| "up"
-| "label"
-| "image"
-| "home"
-| "filter"
-| "expand"
-| "edit"
-| "download"
-| "compress"
-| "close"
-| "chat"
-| "calendar"
-| "bookmark"
-| "announcement"
-| "add"
-| "alert-message"
-| "money"
-| "circle-x"
-| "check"
-| "circle-check"
-| "trash"
-| "share"
-| "mail"
-| "lock"
+export type Icons =
+    | "chevron"
+    | "arrow"
+    | "warning-circle"
+    | "warning-circle-fill"
+    | "information"
+    | "clock"
+    | "warning"
+    | "notification"
+    | "upload"
+    | "new-folder"
+    | "cloud"
+    | "copy"
+    | "document"
+    | "folder"
+    | "new-file"
+    | "arrow-down"
+    | "arrow-left"
+    | "arrow-right"
+    | "arrow-up"
+    | "back"
+    | "caret-down"
+    | "caret-left"
+    | "caret-right"
+    | "caret-up"
+    | "down"
+    | "full-screen"
+    | "menu"
+    | "more"
+    | "next"
+    | "search"
+    | "settings"
+    | "up"
+    | "label"
+    | "image"
+    | "home"
+    | "filter"
+    | "expand"
+    | "edit"
+    | "download"
+    | "compress"
+    | "close"
+    | "chat"
+    | "calendar"
+    | "bookmark"
+    | "announcement"
+    | "add"
+    | "alert-message"
+    | "money"
+    | "circle-x"
+    | "check"
+    | "circle-check"
+    | "trash"
+    | "share"
+    | "mail"
+    | "lock"
 
 export type AlignItems =
     | 'stretch'
@@ -97,7 +97,7 @@ export type JustifyContent =
     | 'initial'
     | 'inherit';
 
-export type AlignContent = 
+export type AlignContent =
     | 'normal'
     | 'start'
     | 'flex-start'
@@ -131,6 +131,6 @@ export type AlignContent =
     | 'initial'
     | 'inherit'
 
-export type FlexDirection = 'row' |'row-reverse' | 'column' |'column-reverse';
+export type FlexDirection = 'row' | 'row-reverse' | 'column' | 'column-reverse';
 
-export type FlexWrap ='wrap' | 'nowrap' | 'wrap-reverse';
+export type FlexWrap = 'wrap' | 'nowrap' | 'wrap-reverse';
